% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_compute_FD.R
\name{run_compute_FD}
\alias{run_compute_FD}
\title{Run computation of FD}
\usage{
run_compute_FD(newicks1 = NULL, newicks2 = NULL, file1 = NULL, file2 = NULL)
}
\arguments{
\item{newicks1}{Character vector of Newick strings for first tree(s) (optional).}

\item{newicks2}{Character vector of Newick strings for target trees(s)  (optional).}

\item{file1}{Path to a file containing Newick trees for first tree(s) (optional).}

\item{file2}{Path to a file containing Newick trees for target trees(s) (optional).}
}
\value{
Output of computeFD
}
\description{
Run computation of FD
}
