% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_compute_rank.R
\name{run_compute_rank}
\alias{run_compute_rank}
\title{Compute ranks}
\usage{
run_compute_rank(newicks = NULL, file = NULL)
}
\arguments{
\item{newicks}{Character vector of Newick strings (optional)}

\item{file}{Path to a file containing Newick trees (optional)}
}
\value{
Output of computeRank
}
\description{
Compute ranks
}
