% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_compute_similarity.R
\name{run_compute_similarity}
\alias{run_compute_similarity}
\title{Run computation of similarity function}
\usage{
run_compute_similarity(
  newicks1 = NULL,
  newicks2 = NULL,
  file1 = NULL,
  file2 = NULL
)
}
\arguments{
\item{newicks1}{Character vector of Newick strings for first tree(s) (optional).}

\item{newicks2}{Character vector of Newick strings for second trees(s) or target tree (optional).}

\item{file1}{Path to a file containing Newick trees for first tree(s) (optional).}

\item{file2}{Path to a file containing Newick trees for second trees(s) or target tree (optional).}
}
\value{
Output of computeSimilarity
}
\description{
Run computation of similarity function
}
