% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FDRcontrol_search.R
\name{run_FDRcontrol_search}
\alias{run_FDRcontrol_search}
\title{Run FDR controled search}
\usage{
run_FDRcontrol_search(
  newicks = NULL,
  newicks1 = NULL,
  newicks2 = NULL,
  file = NULL,
  file1 = NULL,
  file2 = NULL,
  n1 = NULL,
  q = 0.1,
  random_subsampling = FALSE,
  SPbuilder = "basic_score",
  top_width = NULL,
  bottom_width = 1,
  orientation = "upwards",
  alpha = 0.85,
  q0 = 0.5,
  tau = 0.8,
  Mt = 1,
  rb = NULL
)
}
\arguments{
\item{newicks}{Character vector of Newick strings (optional)}

\item{newicks1}{Character vector of Newick strings for Stable Search (optional)}

\item{newicks2}{Character vector of Newick strings for FDR control Search (optional)}

\item{file}{Path to a file containing Newick trees (optional)}

\item{file1}{Path to a file containing Newick trees for Stable Search (optional)}

\item{file2}{Path to a file containing Newick trees for FDR control Search (optional)}

\item{n1}{Number of trees to be used for Stable Search (optional)}

\item{q}{Numeric value passed to FDR control purposes.}

\item{random_subsampling}{Boolean indicating if the subsampling is to be made random}

\item{SPbuilder}{String indication what method to use to build subposet.}

\item{top_width}{Number of trees of maximal rank for the subposet building}

\item{bottom_width}{Number of trees of rank 1 for the subposet building}

\item{orientation}{Orientation for the building of the subposet}

\item{alpha}{Numeric value passed to Stable Search for stable threshold.}

\item{q0}{Numeric value passed to subposet building process when building upwards (optional)}

\item{tau}{Extra value for subposet building.}

\item{Mt}{Number of maximal trees for basic-bifurcation subposet building}

\item{rb}{Anchor rank for basic-bifurcation subposet building}
}
\value{
A length-1 character vector: the Newick string of the FDR-controlled tree
}
\description{
Run FDR controled search
}
