% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_newSubposet.R
\name{run_NewSubposet}
\alias{run_NewSubposet}
\title{Running a complete analysis of the properties of the subposet constructed from a set of samples}
\usage{
run_NewSubposet(newicks = NULL, file = NULL, Mt, rb)
}
\arguments{
\item{newicks}{Character vector of Newick strings for subPoset Building(optional)}

\item{file}{Path to a file containing Newick trees for subPoset Building (optional)}

\item{Mt}{Numeric value representing number of maximal trees in subPoset}

\item{rb}{rank at which bifurcation in subposet starts}

\item{tau}{Extra value for subposet building.}
}
\value{
Output of computeNewSubposet
}
\description{
Running a complete analysis of the properties of the subposet constructed from a set of samples
}
