% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nullCovering_prob.R
\name{run_nullCovering_prob}
\alias{run_nullCovering_prob}
\title{Computing the probability of correctly identifying C null pairs}
\usage{
run_nullCovering_prob(
  tstar_newick = NULL,
  newicks = NULL,
  newicks1 = NULL,
  newicks2 = NULL,
  tstar_file = NULL,
  file = NULL,
  file1 = NULL,
  file2 = NULL,
  n1 = NULL,
  n2 = NULL,
  random_subsampling = FALSE,
  alpha = 0.85,
  q = 0.1,
  tau = 0.95
)
}
\arguments{
\item{tstar_newick}{Character vector of Newick string for the target tree (optional).}

\item{newicks}{Character vector of Newick strings (optional)}

\item{newicks1}{Character vector of Newick strings for Stable Search (optional)}

\item{newicks2}{Character vector of Newick strings for probability computation (optional)}

\item{tstar_file}{Path to a file containing Newick string for the target tree (optional)}

\item{file}{Path to a file containing Newick trees (optional)}

\item{file1}{Path to a file containing Newick trees for Stable Search (optional)}

\item{file2}{Path to a file containing Newick trees for probability computation (optional)}

\item{n1}{Number of trees to be used for Stable Search (optional)}

\item{n2}{Hypothetical size of the second sample, parameter needed to build the subPoset}

\item{random_subsampling}{Boolean indicating if the subsampling is to be made random}

\item{alpha}{Numeric value passed to Stable Search for stable threshold.}

\item{q}{Numeric value used in FDR control purposes, used in the construction of the subPoset.}

\item{tau}{Extra value for subposet building.}
}
\value{
Output of nullCoveringProb
}
\description{
Computing the probability of correctly identifying C null pairs
}
