% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stable_search.R
\name{run_stable_search}
\alias{run_stable_search}
\title{Run stable search}
\usage{
run_stable_search(newicks = NULL, file = NULL, alpha)
}
\arguments{
\item{newicks}{Character vector of Newick strings (optional)}

\item{file}{Path to a file containing Newick trees (optional)}

\item{alpha}{Numeric value passed to stableSearchRcpp}
}
\value{
Output of stableSearchRcpp
}
\description{
Run stable search
}
