% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_subPosAnalysis.R
\name{run_subPostAnalysis}
\alias{run_subPostAnalysis}
\title{Running a complete analysis of the properties of the subposet constructed from a set of samples}
\usage{
run_subPostAnalysis(
  tstar_newick = NULL,
  newicks1 = NULL,
  newicks2 = NULL,
  bnewicks = NULL,
  tstar_file = NULL,
  file1 = NULL,
  file2 = NULL,
  bfile = NULL,
  SPbuilder = "stability",
  alpha = 0.85,
  q = 0.1,
  tau = 0.8,
  Mt = 1,
  rb = NULL,
  delta = NULL
)
}
\arguments{
\item{tstar_newick}{Character vector of Newick string for the target tree (optional).}

\item{newicks1}{Character vector of Newick strings for Stable Search (optional)}

\item{newicks2}{Character vector of Newick strings for subposet Variability Analysis (optional)}

\item{bnewicks}{Character vector of Newick strings for probability computation (optional)}

\item{tstar_file}{Path to a file containing Newick string for the target tree (optional)}

\item{file1}{Path to a file containing Newick trees for Stable Search (optional)}

\item{file2}{Path to a file containing Newick trees for subposet Variability Analysis (optional)}

\item{bfile}{Path to a file containing Newick trees for probability computation (optional)}

\item{SPbuilder}{String indication what method to use to build subposet.}

\item{alpha}{Numeric value passed to Stable Search for stable threshold. (in stability-based SubPoset)}

\item{q}{Numeric value used in FDR control purposes, used in the construction of the subPoset in stability-based subposet}

\item{tau}{Extra value for subposet building in stability-based subposet}

\item{Mt}{Number of maximal trees for basic-bifurcation subposet building}

\item{rb}{Anchor rank for basic-bifurcation subposet building}
}
\value{
Output of SPAnalysisR
}
\description{
Running a complete analysis of the properties of the subposet constructed from a set of samples
}
