% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_FDRcontrol_search.R
\name{run_FDRcontrol_search}
\alias{run_FDRcontrol_search}
\title{Run stable search}
\usage{
run_FDRcontrol_search(
  newicks = NULL,
  newicks1 = NULL,
  newicks2 = NULL,
  file = NULL,
  file1 = NULL,
  file2 = NULL,
  n1 = NULL,
  random_subsampling = FALSE,
  alpha = 0.85,
  q = 0.1,
  tau = 0.95,
  summarized = FALSE
)
}
\arguments{
\item{newicks}{Character vector of Newick strings (optional)}

\item{newicks1}{Character vector of Newick strings for Stable Search (optional)}

\item{newicks2}{Character vector of Newick strings for FDR control Search (optional)}

\item{file}{Path to a file containing Newick trees (optional)}

\item{file1}{Path to a file containing Newick trees for Stable Search (optional)}

\item{file2}{Path to a file containing Newick trees for FDR control Search (optional)}

\item{n1}{Number of trees to be used for Stable Search (optional)}

\item{random_subsampling}{Boolean indicating if the subsampling is to be made random}

\item{alpha}{Numeric value passed to Stable Search for stable threshold.}

\item{q}{Numeric value passed to FDR control purposes.}

\item{tau}{Extra value for subposet building.}

\item{summarized}{Boolean factor indicating if the function is to be runned with a summarized version of the sample}
}
\value{
Output of completeSearchRcpp
}
\description{
Run stable search
}
