% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_subPosAnalysis.R
\name{run_subPostAnalysis}
\alias{run_subPostAnalysis}
\title{Running a complete analysis of the properties of the subposet constructed from a set of samples}
\usage{
run_subPostAnalysis(
  tstar_newick = NULL,
  newicks1 = NULL,
  newicks2 = NULL,
  bnewicks = NULL,
  tstar_file = NULL,
  file1 = NULL,
  file2 = NULL,
  bfile = NULL,
  alpha = 0.85,
  q = 0.1,
  tau = 0.95,
  summarized = FALSE
)
}
\arguments{
\item{tstar_newick}{Character vector of Newick string for the target tree (optional).}

\item{newicks1}{Character vector of Newick strings for Stable Search (optional)}

\item{newicks2}{Character vector of Newick strings for subposet Variability Analysis (optional)}

\item{bnewicks}{Character vector of Newick strings for probability computation (optional)}

\item{tstar_file}{Path to a file containing Newick string for the target tree (optional)}

\item{file1}{Path to a file containing Newick trees for Stable Search (optional)}

\item{file2}{Path to a file containing Newick trees for subposet Variability Analysis (optional)}

\item{bfile}{Path to a file containing Newick trees for probability computation (optional)}

\item{alpha}{Numeric value passed to Stable Search for stable threshold.}

\item{q}{Numeric value used in FDR control purposes, used in the construction of the subPoset.}

\item{tau}{Extra value for subposet building.}

\item{summarized}{Boolean factor indicating if the function is to be runned with a summarized version of the sample}
}
\value{
Output of SPAnalysisR
}
\description{
Running a complete analysis of the properties of the subposet constructed from a set of samples
}
